Separate out python code into multiple files and update gitignore
