+++ src/consts.py
@@ -1,0 +1,22 @@
+# Constants for Total Serialism Gui
+
+TIME_SIGNATURES = [  "4/4",
+            "1/4","3/8", "2/4", "5/8", "6/8",
+            "3/4", "7/8", "9/8", "5/4", "11/8",
+             "12/8", "6/4", "3/2", "13/8",
+            "7/4", "15/8",
+            ]
+
+SINGLE_CLEF = ["violin", "viola", "cello", "bass",
+        "oboe", "bassoon", "clarinet", "piccolo", "flute", 
+        "trombone", "trumpet", "frenchhorn", "tuba"]
+
+MULTI_CLEF = ["piano", "harp"]
+
+PARAM_NAMES = ["Pitch Row", "Rhythm Row", "Articulation Row",
+               "Tempo", "Time Signature", "Title", "Composer", "Number of Rows"]
+
+INSTRUMENT_FIELDS = ["Instrument Name", "Row Numbers", "Row Types", "Dynamics Row"]
+
+MULTI_CLEF_FIELDS = ["Instrument Name", "Right Hand Row Numbers", "Right Hand Row Types", 
+                     "Left Hand Row Numbers", "Left Hand Row Numbers", "Dynamics Row"]
